fix : only clubLeader can delete the club

+++ src/club/club.service.ts
@@ -108,10 +108,13 @@ export class ClubService {
     return ClubDto.from(updatedClub);
   }
 
-  async deleteClub(clubId: number): Promise<void> {
+  async deleteClub(clubId: number, user: UserBaseInfo): Promise<void> {
     const club = await this.clubRepository.getClubById(clubId);
     if (!club) {
       throw new NotFoundException('해당 클럽이 존재하지 않습니다.');
+    }
+    if (club.leaderId !== user.id) {
+      throw new ConflictException('클럽 리더만 수정할 수 있습니다.');
     }
 
     await this.clubRepository.deleteClub(clubId);
